requirements.txt
gradio==5.4.0
python-dotenv==1.0.1
